requirements.txt
Django
djangorestframework
django-cors-headers
djangorestframework-simplejwt
drf-spectacular
pillow
django-filter
requests